# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NFL Auction Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mtm:
    get:
      operationId: getMtmSnapshots
      tags: [mtm]
      summary: Get weekly MTM snapshots for a season
      parameters:
        - name: season
          in: query
          required: true
          schema:
            type: number
        - name: calcuttaId
          in: query
          schema:
            type: ["number", "null"]
        - name: teamId
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: MTM snapshots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MtmData"
    post:
      operationId: upsertMtmSnapshot
      tags: [mtm]
      summary: Add or update a weekly MTM snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MtmSnapshotInput"
      responses:
        "200":
          description: Upserted snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MtmSnapshot"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized — ADMIN_API_KEY required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The target team/date is protected Kalshi Week 0 data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /mtm/week-zero/capture:
    post:
      operationId: captureWeekZeroMtm
      tags: [mtm]
      summary: Capture an admin-authorized Week 0 valuation from public Kalshi market data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeekZeroCaptureInput"
      responses:
        "200":
          description: Captured Week 0 valuation diagnostics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekZeroCaptureResult"
        "400":
          description: Validation or incomplete market data error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized — ADMIN_API_KEY required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Season not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The selected Week 0 date is already occupied by manual MTM data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Kalshi market data could not be captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      description: Administrative bearer token supplied through ADMIN_API_KEY.
    JobRunnerBearerAuth:
      type: http
      scheme: bearer
      description: Restricted job trigger token supplied through JOB_RUNNER_SECRET.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    MtmSnapshot:
      type: object
      properties:
        id:
          type: number
        teamId:
          type: number
        seasonId:
          type: number
        weekNum:
          type: ["number", "null"]
        snapshotDate:
          type: string
        mtmValue:
          type: number
        snapshotKey:
          type: ["string", "null"]
        source:
          type: string
        capturedAt:
          type: ["string", "null"]
        marketStatus:
          type: ["string", "null"]
          enum: [live, stale, incomplete, manual, null]
        bankedPoints:
          type: ["number", "null"]
        seasonEquityPoints:
          type: ["number", "null"]
        bonusEquityPoints:
          type: ["number", "null"]
        totalPoints:
          type: ["number", "null"]
        normalizedShare:
          type: ["number", "null"]
      required: [id, teamId, seasonId, snapshotDate, mtmValue]
    MtmWeekData:
      type: object
      properties:
        snapshotDate:
          type: string
        weekNum:
          type: ["number", "null"]
        label:
          type: string
        source:
          type: string
        capturedAt:
          type: ["string", "null"]
        potSize:
          type: number
        rawPointTotal:
          type: number
        normalizedShareTotal:
          type: number
        marketStatusCounts:
          $ref: "#/components/schemas/MtmMarketStatusCounts"
        ownerTotals:
          type: array
          items:
            $ref: "#/components/schemas/MtmOwnerWeek"
        teamValues:
          type: array
          items:
            $ref: "#/components/schemas/MtmTeamWeek"
      required:
        [
          snapshotDate,
          label,
          source,
          potSize,
          rawPointTotal,
          normalizedShareTotal,
          marketStatusCounts,
          ownerTotals,
          teamValues,
        ]
    MtmMarketStatusCounts:
      type: object
      properties:
        live:
          type: number
        stale:
          type: number
        incomplete:
          type: number
        manual:
          type: number
      required: [live, stale, incomplete, manual]
    MtmOwnerWeek:
      type: object
      properties:
        bidderName:
          type: string
        mtmTotal:
          type: number
      required: [bidderName, mtmTotal]
    MtmTeamWeek:
      type: object
      properties:
        teamId:
          type: number
        teamName:
          type: string
        ownerName:
          type: string
        mtmValue:
          type: number
        source:
          type: string
        capturedAt:
          type: ["string", "null"]
        marketStatus:
          type: string
          enum: [live, stale, incomplete, manual]
        contractSetId:
          type: ["string", "null"]
        marketStatusReasons:
          type: array
          items:
            type: string
        bankedPoints:
          type: ["number", "null"]
        seasonEquityPoints:
          type: ["number", "null"]
        bonusEquityPoints:
          type: ["number", "null"]
        totalPoints:
          type: ["number", "null"]
        normalizedShare:
          type: ["number", "null"]
        winTotalLine:
          type: ["number", "null"]
        winTotalOverProbability:
          type: ["number", "null"]
        rawExpectedWins:
          type: ["number", "null"]
        expectedWins:
          type: ["number", "null"]
        playoffProbability:
          type: ["number", "null"]
        divisionalProbability:
          type: ["number", "null"]
        conferenceGameProbability:
          type: ["number", "null"]
        superBowlProbability:
          type: ["number", "null"]
        championshipProbability:
          type: ["number", "null"]
        regularSeasonMethod:
          type: ["string", "null"]
        intermediateRoundMethod:
          type: ["string", "null"]
        marketQuotes:
          type: array
          items:
            $ref: "#/components/schemas/MtmMarketQuote"
      required:
        [
          teamId,
          teamName,
          ownerName,
          mtmValue,
          source,
          marketStatus,
          marketStatusReasons,
          marketQuotes,
        ]
    MtmMarketQuote:
      type: object
      properties:
        ticker:
          type: string
        kind:
          type: string
          enum: [win_threshold, playoff, conference_champion, championship]
        line:
          type: ["number", "null"]
        bid:
          type: ["number", "null"]
        ask:
          type: ["number", "null"]
        last:
          type: ["number", "null"]
        probability:
          type: ["number", "null"]
        spread:
          type: ["number", "null"]
        bidDepth:
          type: ["number", "null"]
        askDepth:
          type: ["number", "null"]
        quality:
          type: string
          enum: [live, stale, unavailable]
        updatedAt:
          type: ["string", "null"]
        derived:
          type: boolean
      required:
        [
          ticker,
          kind,
          line,
          bid,
          ask,
          last,
          probability,
          spread,
          bidDepth,
          askDepth,
          quality,
          updatedAt,
          derived,
        ]
    MtmData:
      type: object
      properties:
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/MtmWeekData"
        teams:
          type: array
          items:
            $ref: "#/components/schemas/MtmTeamSeries"
        owners:
          type: array
          items:
            $ref: "#/components/schemas/MtmOwnerSeries"
      required: [weeks, teams, owners]
    MtmTeamSeries:
      type: object
      properties:
        teamId:
          type: number
        teamName:
          type: string
        conference:
          type: string
        ownerName:
          type: string
        weeklyValues:
          type: array
          items:
            type: number
      required: [teamId, teamName, conference, ownerName, weeklyValues]
    MtmOwnerSeries:
      type: object
      properties:
        bidderName:
          type: string
        weeklyTotals:
          type: array
          items:
            type: number
      required: [bidderName, weeklyTotals]
    MtmSnapshotInput:
      type: object
      properties:
        teamId:
          type: number
        seasonYear:
          type: number
        calcuttaId:
          type: number
        weekNum:
          type: number
          description: Optional week label (0=pre-season, 1–18=regular, 19+=playoffs)
        snapshotDate:
          type: string
          description: "Date as YYYY-MM-DD. Defaults to today. Same-day submissions overwrite the previous value."
        mtmValue:
          type: number
      required: [teamId, seasonYear, mtmValue]
    WeekZeroCaptureInput:
      type: object
      properties:
        seasonYear:
          type: number
        calcuttaId:
          type: number
        snapshotDate:
          type: string
          description: Optional Week 0 calendar date as YYYY-MM-DD. The first successful capture fixes this date for idempotent retries.
      required: [seasonYear]
    WeekZeroCaptureResult:
      type: object
      properties:
        seasonYear:
          type: number
        snapshotDate:
          type: string
        capturedAt:
          type: string
        teamCount:
          type: number
        contractSetId:
          type: string
        potSize:
          type: number
        rawPointTotal:
          type: number
        normalizedShareTotal:
          type: number
        marketStatusCounts:
          $ref: "#/components/schemas/MtmMarketStatusCounts"
      required:
        [
          seasonYear,
          snapshotDate,
          capturedAt,
          teamCount,
          contractSetId,
          potSize,
          rawPointTotal,
          normalizedShareTotal,
          marketStatusCounts,
        ]

    # Summary
...